+++ docs/changelog.md
@@ -18,6 +18,7 @@ Changelog format adheres to [Keep a Changelog](http://keepachangelog.com/)</sub>
 ### Added
 
 - Updated deps.
+- C2x2 matrix space alias.
 
 ## [v0.12.3] - 2018-03-29